lima: add prime import support
This enables lima to import a shared buffer from a dma_buf handle using
the DRM prime infrastructure.
With this, it is possible to allocate a dumb buffer with a display
driver to render to a display.
We can mostly use helper DRM callbacks, but gem_prime_import_sg_table
must have a custom function because we must allocate a lima_bo structure
and initialize custom fields other than the drm_gem_object.

This was tested on the Allwinner A20 with the sun4i-drm driver, along
with lima.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>